Add ranking behavior to other presenters + introduce order value to keep order in final render
